chore: renames inconsistent filename
