+++ application/next-frontend/src/Landing/scenarios/FirstSection/components/MadeByBlank.tsx
