+++ application/next-frontend/src/Landing/scenarios/FirstSection/components/QuoteBubbles.tsx
