+++ application/next-frontend/src/Landing/scenarios/FirstSection/index.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image'
 import { AddToSlackButtonWhiteRounded } from '@/Landing/components/AddToSlackButtons'
 import Mascot from '@/Landing/assets/illustrations/MascotWithFourArmsGreen.svg'
 import SpeachBubble from '@/Landing/assets/illustrations/SpeachBubble.svg'
-import { MadeByBlank } from './components/madeByBlank'
-import { QuoteBubbles } from './components/quotesBubbles'
+import { MadeByBlank } from './components/MadeByBlank'
+import { QuoteBubbles } from './components/QuoteBubbles'
 
 const RightSideMascot = () => {
     return (
